WIP: Split out loop logic from Pregel class
