+++ libs/langgraph/langgraph/pregel/types.py
@@ -88,3 +88,12 @@ class StateSnapshot(NamedTuple):
 
 
 All = Literal["*"]
+
+StreamMode = Literal["values", "updates", "debug"]
+"""How the stream method should emit outputs.
+
+- 'values': Emit all values of the state for each step.
+- 'updates': Emit only the node name(s) and updates
+    that were returned by the node(s) **after** each step.
+- 'debug': Emit debug events for each step.
+"""
