+++ libs/langgraph/tests/test_pregel_async.py
@@ -623,8 +623,6 @@ async def test_invoke_two_processes_in_out(mocker: MockerFixture) -> None:
 
     assert await app.ainvoke(2) == 4
 
-    assert await app.ainvoke(2, input_keys="inbox") == 3
-
     with pytest.raises(GraphRecursionError):
         await app.ainvoke(2, {"recursion_limit": 1})
